# mp-app/openapi.yaml
components:
  schemas:
...
    AvailabilityUpdate:
      additionalProperties: false
      properties:
        available:
          title: Available
          type: boolean
      required:
      - available
      title: AvailabilityUpdate
      type: object
...
    EquipmentCreate:
      additionalProperties: false
      properties:
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        count:
          default: 1
          minimum: 1.0
          title: Count
          type: integer
        name:
          maxLength: 120
          minLength: 1
          title: Name
          type: string
      required:
      - name
      title: EquipmentCreate
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    MemberCreate:
      additionalProperties: false
      properties:
        birth_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Birth Date
        first_name:
          maxLength: 80
          minLength: 1
          title: First Name
          type: string
        guardian_consent:
          default: false
          title: Guardian Consent
          type: boolean
        last_name:
          maxLength: 80
          minLength: 1
          title: Last Name
          type: string
        membership_code:
          maxLength: 40
          minLength: 1
          title: Membership Code
          type: string
        membership_exp:
          anyOf:
          - type: string
          - type: 'null'
          title: Membership Exp
        phone:
          anyOf:
          - maxLength: 32
            type: string
          - type: 'null'
          title: Phone
        sex:
          enum:
          - male
          - female
          title: Sex
          type: string
      required:
      - membership_code
      - first_name
      - last_name
      - sex
      title: MemberCreate
      type: object
    MemberOut:
      description: Studio view — full record, no clinician notes (those live on injuries).
      properties:
        active_injuries:
          default: 0
          title: Active Injuries
          type: integer
        birth_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Birth Date
        created_at:
          title: Created At
          type: string
        first_name:
          title: First Name
          type: string
        guardian_consent:
          title: Guardian Consent
          type: boolean
        id:
          title: Id
          type: integer
        last_name:
          title: Last Name
          type: string
        membership_code:
          title: Membership Code
          type: string
        membership_exp:
          anyOf:
          - type: string
          - type: 'null'
          title: Membership Exp
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        sex:
          enum:
          - male
          - female
          title: Sex
          type: string
      required:
      - id
      - membership_code
      - first_name
      - last_name
      - sex
      - birth_date
      - phone
      - membership_exp
      - guardian_consent
      - created_at
      title: MemberOut
      type: object
...
    MemberUpdate:
      additionalProperties: false
      properties:
        birth_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Birth Date
        first_name:
          anyOf:
          - maxLength: 80
            minLength: 1
            type: string
          - type: 'null'
          title: First Name
        guardian_consent:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Guardian Consent
        last_name:
          anyOf:
          - maxLength: 80
            minLength: 1
            type: string
          - type: 'null'
          title: Last Name
        membership_exp:
          anyOf:
          - type: string
          - type: 'null'
          title: Membership Exp
        phone:
          anyOf:
          - maxLength: 32
            type: string
          - type: 'null'
          title: Phone
        sex:
          anyOf:
          - enum:
            - male
            - female
            type: string
          - type: 'null'
          title: Sex
      title: MemberUpdate
      type: object
    NutritionRequest:
      additionalProperties: false
      properties:
        activity:
          default: moderate
          enum:
          - sedentary
          - light
          - moderate
          - active
          - athlete
          title: Activity
          type: string
        goal:
          default: maintain
          enum:
          - cut
          - maintain
          - bulk
          title: Goal
          type: string
        protein_g_per_kg:
          default: 1.8
          exclusiveMinimum: 0.0
          maximum: 4.0
          title: Protein G Per Kg
          type: number
      title: NutritionRequest
      type: object
...
    SetMemberPin:
      additionalProperties: false
      properties:
        pin:
          maxLength: 32
          minLength: 4
          title: Pin
          type: string
      required:
      - pin
      title: SetMemberPin
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: Local-first Gym OS core API (offline-capable).
  title: Muscle Paradise Core
  version: 0.20.0
  x-generated-by: app.export_openapi
openapi: 3.1.0
paths:
...
  /api/v1/equipment:
    get:
      operationId: list_all_api_v1_equipment_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                title: Response List All Api V1 Equipment Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: All equipment
      tags:
      - equipment
    post:
      operationId: create_api_v1_equipment_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Api V1 Equipment Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add equipment
      tags:
      - equipment
  /api/v1/equipment/{equipment_id}:
    patch:
      operationId: set_availability_api_v1_equipment__equipment_id__patch
      parameters:
      - in: path
        name: equipment_id
        required: true
        schema:
          title: Equipment Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Set Availability Api V1 Equipment  Equipment Id  Patch
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Toggle availability
      tags:
      - equipment
...
  /api/v1/members:
    get:
      description: OWNER/ADMIN/RECEPTION see everyone; TRAINER only assigned members.
      operationId: list_members_api_v1_members_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MemberOut'
                title: Response List Members Api V1 Members Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List visible members
      tags:
      - members
    post:
      operationId: create_member_api_v1_members_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Register a member
      tags:
      - members
  /api/v1/members/{member_id}:
    delete:
      operationId: delete_member_api_v1_members__member_id__delete
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Soft-delete (tombstone) a member
      tags:
      - members
    get:
      operationId: get_member_api_v1_members__member_id__get
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: One member
      tags:
      - members
    patch:
      operationId: update_member_api_v1_members__member_id__patch
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update a member
      tags:
      - members
...
  /api/v1/members/{member_id}/pin:
    post:
      description: 'Store a PBKDF2 hash of the member''s PIN so they can use the client app.


        Only the hash is written; the plaintext never touches disk (C11).'
      operationId: set_member_pin_api_v1_members__member_id__pin_post
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetMemberPin'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Set member PIN (front desk)
      tags:
      - members
...
  /api/v1/members/{member_id}/qr:
    get:
      description: 'QR payload per map §8: ``{v, typ, gym, mid, exp, sig}`` HMAC-signed.'
      operationId: member_qr_api_v1_members__member_id__qr_get
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Member Qr Api V1 Members  Member Id  Qr Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Signed short-lived check-in QR payload
      tags:
      - members
  /api/v1/nutrition/members/{member_id}/plan:
    get:
      operationId: get_plan_api_v1_nutrition_members__member_id__plan_get
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Plan Api V1 Nutrition Members  Member Id  Plan Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Latest stored nutrition plan
      tags:
      - nutrition
    post:
      operationId: create_plan_api_v1_nutrition_members__member_id__plan_post
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NutritionRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Plan Api V1 Nutrition Members  Member Id  Plan Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Compute + store a nutrition plan from the latest assessment LBM
      tags:
      - nutrition
...